pyproject.toml
[project]
name = "ai-design"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.2.0",
    "notebook>=7.5.4",
    "openai>=2.15.0",
    "pyaudio>=0.2.14",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "sounddevice>=0.5.3",
    "soundfile>=0.13.1",
    "datasets>=3.0.0",
    "jinja2>=3.1.0",
    "kaggle>=1.6.0",
    "scikit-learn>=1.8.0",
    "matplotlib>=3.10.8",
]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
